+++ CHANGELOG.md
@@ -3,3 +3,4 @@
 * add `SparseMatrix.min` and `SparseMatrix.max`
 * `SparseMatrix.getitem` will raise IndexError on invalid coordinates
 * added `index_of` and `index_of_max`
+* added `exhaust` to `is_empty`